Compute the grid Hessian in real precision with a mixed output buffer
The grid Hessian (gridHessian.cu + gridHessianTiled.cu) computed its value,
gradient, and second-derivative accumulators in float and wrote them to a float
buffer, so a double context narrowed the Hessian to fp32 -- the normal-mode /
entropy inputs lose precision exactly in high-curvature clash regions. Template
the Hessian chain-rule helpers (applyHessianChainRule, applyRuntimeCapHessian
ChainRule[Tiled], applyThirdDerivChainRule) on the scalar type and convert the
kernel-body accumulators to real (the 216-coefficient assembly stays double).
Make hessianBuffer/thirdDerivBuffer mixed: the kernels write real, the host
allocates and reads them back at the context precision into double members, and
getHessianBlocks returns double values. analyzeHessianKernel's block parameter
becomes mixed* to match (its eigendecomposition stays float for now -- a separate
axis). __global__ param lists stay float for the host ABI; effectiveCap is real so
the templated cap helper deduces consistently.

single==mixed (per-atom blocks, no cross-atom summation, double assembly in both);
double agrees ~1.7e-6 and now carries the double-precision Hessian. Existing
GBSA/isolated/bonded Hessian regressions unchanged.

+++ platforms/cuda/src/CudaGridForceKernels.h
@@ -244,12 +244,12 @@ private:
     CUfunction hessianKernel;                    // Kernel for Hessian computation
     CUfunction tiledHessianKernel;               // Kernel for tiled Hessian computation
     OpenMM::CudaArray hessianBuffer;            // Per-atom Hessian storage (6 floats per atom)
-    std::vector<float> lastHessianBlocks;        // Persistent copy of last Hessian computation
+    std::vector<double> lastHessianBlocks;        // Persistent copy of last Hessian computation
 
     // Third derivative computation support (method 4 only)
     CUfunction thirdDerivKernel;                 // Kernel for third derivative computation
     OpenMM::CudaArray thirdDerivBuffer;         // Per-atom storage (10 floats per atom)
-    std::vector<float> lastThirdDerivBlocks;     // Persistent copy of last computation
+    std::vector<double> lastThirdDerivBlocks;     // Persistent copy of last computation
 
     // Hessian analysis support (eigendecomposition, curvature, entropy)
     CUfunction analysisKernel;                   // Kernel for per-atom Hessian analysis
